+++ client/src/components/AppLayout.tsx
@@ -137,7 +137,7 @@ export default function AppLayout({ children }: AppLayoutProps) {
 
         {/* Content Area */}
         {isOrderDetail ? (
-          <div className="flex-1 min-h-0 overflow-hidden flex flex-col">
+          <div className="flex-1 flex flex-col min-h-0">
             {children}
           </div>
         ) : (